Dear reader,
I'm sorry you had to go through that code.

Believe me, it was already a mess when I started this commit.
I only tried to change it to meet the requirements and keeping so that
the current maintainers can continue to work on it.

I don't even think we can make an excuse for this... Keeping it this way
and writing more spaghetti code hurt me as well.

I'm so sorry.

Cheers,
frm